bsdinstall: add a dialog for network proxy setup
This adds the possibility in the installer to setup an HTTP, FTP, and/or
SOCKS5 network proxy, as well as a list of exceptions. This is
implemented within a single dialog, once the networking setup is
performed. The default action confirms the setup if any proxy is already
set, or skips this step otherwise.

The default value for the HTTP proxy is obtained from bsdconfig's
`httpProxy` variable; when not set, the default value for this setting
as well as for the FTP and SOCKS5 proxies and the list of exceptions are
obtained from the corresponding environment variable. (If set)

When confirmed, these values are exported back to the environment, as
expected by fetch(3) in the subsequent `distfetch` phase.

PR:		214390
Sponsored By:	The FreeBSD Foundation